requirements.txt
django
psycopg2
djangorestframework
dj-rest-auth
django-filters